Add files via upload
C++ code that goes through a Binary Search Tree and deletes every leaf node with an odd key value. The function should display the key values of each node of the tree, starting with the left child, then the right child, before displaying the root of each subtree. The nodes that are being deleted should display the key with a message “deleting” next to the key value.  The code should prevent any potential memory leaks.